--- Models/LoginModel.cs
@@ -1,9 +1,0 @@
-namespace Models
-{
-    public class LoginModel
-    {
-        public string Email { get; set; } = null!;
-
-        public string Password { get; set; } = null!;
-    }
-}
