+++ Modules/Auth/AuthModel.cs
@@ -1,11 +1,10 @@
-namespace Modules.Auth
+namespace Modules.Auth.Model;
+
+public class AuthModel
 {
-    public class AuthModel
-    {
-        public required string Name { get; set; }
+    public required string Name { get; set; }
 
-        public required string Role { get; set; }
+    public required string Role { get; set; }
 
-        public required Guid Id { get; set; }
-    }
+    public required Guid Id { get; set; }
 }
