+++ Modules/Auth/AuthRequiredAttribute.cs
@@ -1,6 +1,7 @@
 using System.Security.Claims;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Filters;
+using Modules.Auth.Service;
 
 namespace Modules.Auth;
