+++ Modules/Auth/AuthService.cs
@@ -1,21 +1,21 @@
-using Modules.User;
+using Modules.Auth.Model;
+using Modules.User.Service;
 
-namespace Modules.Auth
+namespace Modules.Auth.Service;
+
+public class AuthService(IHttpContextAccessor httpContextAccessor, UserService userService)
 {
-    public class AuthService(IHttpContextAccessor httpContextAccessor, UserService userService)
+    private readonly IHttpContextAccessor _httpContextAccessor = httpContextAccessor;
+    private readonly UserService _userService = userService;
+
+    public async Task<AuthModel?> GetAuthUser()
     {
-        private readonly IHttpContextAccessor _httpContextAccessor = httpContextAccessor;
-        private readonly UserService _userService = userService;
+        var authHeader =
+            _httpContextAccessor.HttpContext?.Request.Headers.Authorization.FirstOrDefault();
+        if (authHeader == null || !authHeader.StartsWith("Bearer "))
+            return null;
 
-        public async Task<AuthModel?> GetAuthUser()
-        {
-            var authHeader =
-                _httpContextAccessor.HttpContext?.Request.Headers.Authorization.FirstOrDefault();
-            if (authHeader == null || !authHeader.StartsWith("Bearer "))
-                return null;
-
-            var token = authHeader["Bearer ".Length..].Trim();
-            return await _userService.Me(token);
-        }
+        var token = authHeader["Bearer ".Length..].Trim();
+        return await _userService.Me(token);
     }
 }
