+++ Modules/Comments/CommentModule.cs
@@ -1,11 +1,14 @@
-namespace Modules.Comments
+using Modules.Comments.Controller;
+using Modules.Comments.Service;
+
+namespace Modules.Comments.Module
 {
-    public static class CommentModule
+    public static class CommentsModule
     {
         public static IServiceCollection AddCommentModule(this IServiceCollection services)
         {
-            services.AddScoped<CommentService>();
-            services.AddControllers().AddApplicationPart(typeof(CommentController).Assembly);
+            services.AddScoped<CommentsService>();
+            services.AddControllers().AddApplicationPart(typeof(CommentsController).Assembly);
             return services;
         }
     }
